+++ Command/BuildCommand.php
@@ -43,30 +43,23 @@ class BuildCommand extends Command
         $this->setName('massive:build');
         $this->setDescription('Execute build or build targets');
         $this->setHelp(<<<EOT
-The <info>%command.full_name%</info> launches a specific build target if specified or the 
-entire build if no arguments are given:
-
-    <info>$ php app/console %command.full_name%</info>
-
-Launch a specific target:
+Build a target:
 
     <info>$ php app/console %command.full_name% mytarget</info>
 
-If you want to see which targets are available, use the <comment>--nobuild</comment> option:
+If you want to see which targets are available run the command without any arguments:
 
-    <info>$ php app/console %command.full_name% --nobuild</info>
+    <info>$ php app/console %command.full_name%</info>
 
-By default if you specify a specific target, the build system will build any dependencies
-it has, to disable this use the <comment>--nodeps</comment> option.
+By default the build system will build any dependencies that the target
+has, to disable this use the <comment>--nodeps</comment> option.
 
     <info>$ php app/console %command.full_name% --nodeps</info>
 EOT
         );
 
-
-        $this->addArgument('target', InputArgument::OPTIONAL, 'Build the specified target', null);
+        $this->addArgument('target', InputArgument::OPTIONAL, 'Target to build', null);
         $this->addOption('nodeps', 'D', InputOption::VALUE_NONE, 'Ignore dependencies');
-        $this->addOption('nobuild', null, InputOption::VALUE_NONE, 'Do not build, only list the targets');
     }
 
     /**
@@ -90,7 +83,7 @@ EOT
 
         $this->renderTargets($builders);
 
-        if ($input->getOption('nobuild')) {
+        if ($target === null) {
             return 0;
         }