+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 Changelog
 =========
 
-0.2
----
+dev-master
+----------
 
+- [feature] Target is mandatory (no longer possible to build all targets)
 - [feature] Allow creation of virtual targets via. configuration
 - [bug] Dependency not found when specifying target (nested dependencies not taken into account)